+++ packages/core/client/src/schema-component/antd/grid-card/GridCard.tsx
@@ -89,6 +89,7 @@ const InternalGridCard = (props: GridCardProps) => {
   const field = useField<ArrayField>();
   const Designer = useDesigner();
   const height = useGridCardBodyHeight();
+  console.log(fieldSchema);
   const [schemaMap] = useState(new Map());
   const getSchema = useCallback(
     (key) => {
